requirements.txt
alembic==1.10.3
click==7.1.2
Flask==1.1.4
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.4
MarkupSafe==2.0.1
python-dotenv==0.17.1
SQLAlchemy==1.4.47
typing_extensions==4.5.0
Werkzeug==0.16.1
WTForms==2.3.3
Flask-WTF==0.15.1
email-validator==1.1.2